License project under Creative Commons Zero v1.0 Universal
It is important to define the license under which a software project is released in order to allow its use.

Previously, although the licensing was indicated by informal comments in most of the sketches, a formal license was not
provided.

This project is not truly a library, but rather just a collection of sketches packaged as a library in order to allow
distribution via Library Manager, but without any library source code. So it is those sketches that are the significant
content of the project, for which licensing must be defined.

As is standard practice for the example sketches created by Arduino, this project's sketches contain the statement:

> This example code is in the public domain.

The appropriate formal license in this case is Creative Commons Zero v1.0 Universal.

A standardized Creative Commons Zero v1.0 Universal license file is hereby added to the project. In addition to
providing this important information to the project users, this will also allow automated license type detection (e.g.,
by the Licensee Gem used by the GitHub website).